Adding employment question navigation with updates
